moved client out of example allowing ext usage
